+++ Everest.ViewModels/Request/CreacionAnuncioRequest.cs
@@ -3,7 +3,7 @@
     public class CreacionAnuncioRequest
     {
         public int IdUsuario { get; set; }
-        public decimal Metros2 { get; set; }
+        public int Metros2 { get; set; }
         public int CantidadHabitaciones { get; set; }
         public int CantidadBaños { get; set; }
         public int CantidadParqueos { get; set; }
